+++ rendering/surfaces/PCDX11RenderTexture.h
@@ -43,7 +43,7 @@ public:
 	void initForRenderTarget(IRenderSurface *renderSurface, uint32_t format, ID3D11Texture2D *texture);
 
 	void resFree() override;
-	void resFill(void* src, size_t size, size_t offset) override;
+	void resFill(void* src, uint32_t size, uint32_t offset) override;
 	char *resGetBuffer() override;
 	void resConstruct() override;
 	// void resMethod10();
